dhruvin: improve mower error checking
This commit add a few changes
- Changed input for finishMove function to Move, lawnmower state and
turns stalled.
The change is to make sure lawnmower is only aware of
information available to it and not more. We were passing co-ordinates
from mowerState.

- Added functionality to finishMove function
Earlier, the function only supported function for "move" case. The
function has been changed to work with stalled, crashed and move cases.

- mower action check has been updated
there were two if statements doing incremental repeated checks. I added
a for loop that can do the same in less space. Also, moved direction
change to the end. The reason for this is to makesure that move is
successful before we change direction. I am also passing no of turns
stalled to the Mower as an added feedback.